% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astrocalc.r
\name{astrocalc}
\alias{astrocalc}
\title{astrocalc}
\usage{
astrocalc(
  day,
  month,
  year,
  hour,
  timezone,
  lat,
  lon,
  withinput = FALSE,
  seaorland = "maritime",
  acknowledgment = FALSE
)
}
\description{
astrocalc
}
