% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWindchill.R
\name{calcWindchill}
\alias{calcWindchill}
\title{Wind chill (US NWS formula)}
\usage{
calcWindchill(airTemp, sfcWind, roundBy = 1)
}
\arguments{
\item{airTemp}{Numeric. Air temperature in degrees Fahrenheit.}

\item{sfcWind}{Numeric. Wind speed in mph.}

\item{roundBy}{Integer. Number of decimal places to round to (default 1).}
}
\value{
Numeric vector of wind chill values.
}
\description{
Computes wind chill using the standard U.S. National Weather Service formula.
}
\examples{
calcWindchill(airTemp = 30, sfcWind = 10)
calcWindchill(airTemp = c(30, 25), sfcWind = c(10, 15), roundBy = 0)

}
