% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnits.R
\name{convertUnits}
\alias{convertUnits}
\title{convertUnits}
\usage{
convertUnits(
  what,
  from,
  to,
  ignoreattr = FALSE,
  quiet = FALSE,
  return_original_as_attr = TRUE
)
}
\description{
convertUnits
}
