% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degtoradian.R
\name{degtoradian}
\alias{degtoradian}
\title{Degrees to radians}
\usage{
degtoradian(deg)
}
\arguments{
\item{deg}{Numeric vector of angles in degrees.}
}
\value{
Numeric vector of angles in radians.
}
\description{
Degrees to radians
}
