% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationpressure.R
\name{stationpressure}
\alias{stationpressure}
\title{stationpressure}
\usage{
stationpressure(
  pressureMB,
  airTemp,
  elevation,
  airTempUnits = "C",
  elevUnits = "m"
)
}
\description{
stationpressure
}
