% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winddeg.R
\name{winddeg}
\alias{winddeg}
\title{Wind direction degrees to 16-point compass labels}
\usage{
winddeg(windDeg)
}
\arguments{
\item{windDeg}{Numeric vector of wind directions in degrees.}
}
\value{
Character vector of compass labels (same length as \code{windDeg}).
}
\description{
Converts wind direction degrees into standard 16-point compass labels:
N, NNE, NE, ENE, E, ESE, SE, SSE, S, SSW, SW, WSW, W, WNW, NW, NNW.
}
\examples{
winddeg(c(0, 20, 45, 90, 200, 359.9))

}
