% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenithinfo.R
\name{zenithinfo}
\alias{zenithinfo}
\title{Solar zenith information}
\usage{
zenithinfo(sun, lon, lat)
}
\arguments{
\item{sun}{Date-time(s) at which to compute solar geometry.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}
}
\value{
A list or data.frame with zenith-related quantities (document what you return).
}
\description{
Solar zenith information
}
