% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationpressure.R
\name{stationpressure}
\alias{stationpressure}
\title{Estimate station pressure from sea-level pressure and elevation}
\usage{
stationpressure(
  pressureMB,
  airTemp,
  elevation,
  airTempUnits = "C",
  elevUnits = "m"
)
}
\arguments{
\item{pressureMB}{Numeric vector of pressure (typically sea-level) in millibars/hPa.}

\item{airTemp}{Numeric vector of air temperature.}

\item{elevation}{Numeric vector of station elevation.}

\item{airTempUnits}{Character temperature units for \code{airTemp} (e.g. \code{"degC"} or \code{"degF"}).}

\item{elevUnits}{Character elevation units for \code{elevation} (e.g. \code{"m"} or \code{"feet"}).}
}
\value{
Numeric vector of estimated station pressure (typically in millibars/hPa).
}
\description{
Estimate station pressure from sea-level pressure and elevation
}
